mqtt_publish/requirements.txt
psycopg2-binary
paho-mqtt
